Carry planet positions in float64 on the CPU; float32 stays the GPU bake
Pluto stepped in 449 km quanta along its own orbit ring — 0.38 of its
radius. positionsAt wrote parsecs into a Float32Array that was not just
the iLocalRel attribute: planetLocalPositionInto and friends read it, so
the mesh LOD, focus path, moving-focal ride, and overlay projections all
inherited the quantum. Neptune shares the binade and therefore the same
449 km step; it reads as smooth only because it is 20x the radius.

PlanetBodyField gains a Float64Array master (localRel64) that every CPU
consumer reads; bufs.localRel is baked from it each write and continues
to feed the GPU billboard, where 449 km is sub-pixel. positionsAt widens
to Float64Array. Moons compose through the same buffer and inherit it.

+++ src/client/stellata.ts
@@ -1181,10 +1181,10 @@ export class Stellata implements FrameAnchor {
   /** Renderer-local positions of the focused host's planets (xyz
    *  triples, length 3·N), or null if no system is attached. Host
    *  offset is applied — under planet focus the host is not at the
-   *  local origin. Returns a fresh Float32Array copy each call (see
+   *  local origin. Returns a fresh Float64Array copy each call (see
    *  `PlanetBodyField.getHostLocalPositions`) — safe to cache across
    *  frames; the value semantics survive attach grow / detach shift. */
-  getFocusedPlanetLocalPositions(): Float32Array | null {
+  getFocusedPlanetLocalPositions(): Float64Array | null {
     const ps = this.focus.getFocusedPlanetSystem();
     if (!ps) return null;
     const rel = this.planetBodyField.getHostLocalPositions(ps.hostStarIdx);
